Refactoring draw plugin into page and pencil
