+++ core/engines/world/client.js
@@ -1631,8 +1631,10 @@ class World {
                 const {name: srcName} = srcItem;
                 const componentApis = tags.getTagComponentApis(srcName);
 
-                const _forEachSrcTagAttribute = fn => {
-                  for (let i = 0; i < componentApis.length; i++) {
+                for (let i = 0; i < componentApis.length; i++) {
+                  const componentApi = componentApis[i];
+
+                  const _forEachSrcTagAttribute = fn => {
                     const componentApi = componentApis[i];
                     const {attributes: componentAttributes = {}} = componentApi;
 
@@ -1645,64 +1647,64 @@ class World {
 
                       fn(attributeName, attributeValue);
                     }
-                  }
-                };
+                  };
 
-                if (!dstTagMesh) {
-                  const {item} = srcTagMesh;
+                  if (!dstTagMesh) {
+                    const {item} = srcTagMesh;
 
-                  const itemSpec = _clone(item);
-                  itemSpec.id = _makeId();
-                  itemSpec.type = 'entity';
-                  const tagName = (() => {
-                    for (let i = 0; i < componentApis.length; i++) {
-                      const componentApi = componentApis[i];
+                    const itemSpec = _clone(item);
+                    itemSpec.id = _makeId();
+                    itemSpec.type = 'entity';
+                    const tagName = (() => {
                       const {selector: componentSelector = 'div'} = componentApi;
                       const {rule: {tagName}} = cssSelectorParser.parse(componentSelector);
 
                       if (tagName) {
                         return tagName;
+                      } else {
+                        return 'entity';
                       }
-                    }
-                    return 'entity';
-                  })();
-                  itemSpec.tagName = tagName;
-                  const attributes = (() => {
-                    const result = {};
-                    _forEachSrcTagAttribute((attributeName, attributeValue) => {
-                      result[attributeName] = {
-                        value: attributeValue,
-                      };
-                    });
-                    return result;
-                  })();
-                  itemSpec.attributes = attributes;
-                  const matrix = (() => {
-                    const {matrix: oldMatrix} = itemSpec;
-                    const position = new THREE.Vector3().fromArray(oldMatrix.slice(0, 3));
-                    const rotation = new THREE.Quaternion().fromArray(oldMatrix.slice(3, 3 + 4));
-                    const scale = new THREE.Vector3().fromArray(oldMatrix.slice(3 + 4, 3 + 4 + 3));
-
-                    position.add(new THREE.Vector3(0, 0, 0.1).applyQuaternion(rotation));
-
-                    return position.toArray().concat(rotation.toArray()).concat(scale.toArray());
-                  })();
-                  itemSpec.matrix = matrix;
-
-                  _addTag(itemSpec, 'world');
-                } else {
-                  const {item: dstItem} = dstTagMesh;
-                  const {id: dstId, instance: dstElement} = dstItem;
-
-                  _forEachSrcTagAttribute((attributeName, attributeValue) => {
-                    if (!dstElement.hasAttribute(attributeName)) {
-                      _setAttribute({
-                        id: dstId,
-                        name: attributeName,
-                        value: attributeValue,
+                    })();
+                    itemSpec.name = tagName;
+                    itemSpec.displayName = tagName;
+                    itemSpec.tagName = tagName;
+                    const attributes = (() => {
+                      const result = {};
+                      _forEachSrcTagAttribute((attributeName, attributeValue) => {
+                        result[attributeName] = {
+                          value: attributeValue,
+                        };
                       });
-                    }
-                  });
+                      return result;
+                    })();
+                    itemSpec.attributes = attributes;
+                    const matrix = (() => {
+                      const {matrix: oldMatrix} = itemSpec;
+                      const position = new THREE.Vector3().fromArray(oldMatrix.slice(0, 3));
+                      const rotation = new THREE.Quaternion().fromArray(oldMatrix.slice(3, 3 + 4));
+                      const scale = new THREE.Vector3().fromArray(oldMatrix.slice(3 + 4, 3 + 4 + 3));
+
+                      position.add(new THREE.Vector3(0, 0, 0.1).applyQuaternion(rotation));
+
+                      return position.toArray().concat(rotation.toArray()).concat(scale.toArray());
+                    })();
+                    itemSpec.matrix = matrix;
+
+                    _addTag(itemSpec, 'world');
+                  } else {
+                    const {item: dstItem} = dstTagMesh;
+                    const {id: dstId, instance: dstElement} = dstItem;
+
+                    _forEachSrcTagAttribute((attributeName, attributeValue) => {
+                      if (!dstElement.hasAttribute(attributeName)) {
+                        _setAttribute({
+                          id: dstId,
+                          name: attributeName,
+                          value: attributeValue,
+                        });
+                      }
+                    });
+                  }
                 }
               }
             }
