+++ plugins/z-draw/package.json
@@ -1,0 +1,11 @@
+{
+  "name": "z-draw",
+  "description": "Paint 2d pixel pictures.",
+  "version": "0.0.1",
+  "keywords": [],
+  "client": "client.js",
+  "server": "server.js",
+  "dependencies": {
+    "mod-loop": "0.0.2"
+  }
+}
