+++ plugins/z-draw/server.js
@@ -1,0 +1,36 @@
+const path = require('path');
+
+class Draw {
+  constructor(archae) {
+    this._archae = archae;
+  }
+
+  mount() {
+    const {_archae: archae} = this;
+    const {express, app} = archae.getCore();
+
+    const brushesStatic = express.static(path.join(__dirname, 'brushes'));
+    function serveBrushes(req, res, next) {
+      brushesStatic(req, res, next);
+    }
+    app.use('/archae/draw/brushes', serveBrushes);
+
+    this._cleanup = () => {
+      function removeMiddlewares(route, i, routes) {
+        if (route.handle.name === 'serveBrushes') {
+          routes.splice(i, 1);
+        }
+        if (route.route) {
+          route.route.stack.forEach(removeMiddlewares);
+        }
+      }
+      app._router.stack.forEach(removeMiddlewares);
+    };
+  }
+
+  unmount() {
+    this._cleanup();
+  }
+}
+
+module.exports = Draw;
